dhcp: Stop randomizing byte order in packets
h.options.SelectOrderOrAll() randomizes the packer
order in a reply when used. This causes the tests
to panic upon seeing an out-of-order sequence of bytes
even with similar looking bytes.

Use h.options.SelectOrder() and only send the bytes
that are set. With this change the order must be set.

Ref: https://github.com/krolaw/dhcp4/blob/ffbbb16f447b3e05b80e8ddecbb8ada36e2ae006/helpers.go#L11

Signed-off-by: Simarpreet Singh <simar@linux.com>